Simplify iterated_code reduction
This update will attempt to simplify the loop that reads the commands in the
initial array in order. If the default "skip" value is replaced by an empty
string, then a conditional check can be removed, improving efficiency.

Again, if this change comes at the cost of functionality or scalability that I
am unaware of, feel free to revert it.